List Available Geocoder Services

Evaluates the logged in user from an authorization token and returns a data.frame containing the available geocoding services for the associated token.

For users who have not signed into a private portal or ArcGIS Online, the public ArcGIS World Geocoder is used. Otherwise, the first available geocoding service associated with your authorization token is used.

list_geocoders(token = arc_token())

default_geocoder(token = arc_token())

Arguments

token

an object of class httr2_token as generated by auth_code() or related function

Value

a data.frame with columns url, northLat, southLat, eastLon, westLon, name, suggest, zoomScale, placefinding, batch.

Details

The default_geocoder() will return the ArcGIS World Geocoder if no token is available. list_geocoder() requires an authorization token.

To manually create a GeocodeServer object, see geocode_server().

Examples

# Default geocoder object
# ArcGIS World Geocoder
default_geocoder()
#> <GeocodeServer>
#> Description: World Geocoder
#> Version: 11.1
#> CRS: 4326

# Requires an Authorization Token
if (FALSE) {
list_geocoders()
}