Storing Geocoding Results

The results of geocoding operations cannot be stored or persisted unless the for_storage argument is set to TRUE. The default argument value is for_storage = FALSE, which indicates the results of the operation can't be stored, but they can be temporarily displayed on a map, for instance. If you store the results, in a database, for example, you need to set this parameter to true.

Details

See the official documentation for more context.