ArcGIS World Geocoder

The ArcGIS World Geocoder is made publicly available for some uses. The world_geocoder object is used as the default GeocodeServer object in default_geocoder() when no authorization token is found. The find_address_candidates(), reverse_geocode(), and suggest_places() can be used without an authorization token. The geocode_addresses() funciton requires an authorization token to be used for batch geocoding.

Usage

world_geocoder

Format

An object of class GeocodeServer (inherits from list) of length 12.

Value

an object of class GeocodeServer