Search Suggestion

This function returns candidate locations based on a partial search query. It is designed to be used in an interactive search experience in a client facing application.

suggest_places(
  text,
  location = NULL,
  category = NULL,
  search_extent = NULL,
  max_suggestions = NULL,
  country_code = NULL,
  preferred_label_values = NULL,
  geocoder = default_geocoder(),
  token = arc_token()
)

Arguments

text

a scalar character of search key to generate a place suggestion.

location

an sfc_POINT object that centers the search. Optional.

category

a scalar character. Place or address type that can be used to filter suggest results. Optional.

search_extent

an object of class bbox that limits the search area. This is especially useful for applications in which a user will search for places and addresses within the current map extent. Optional.

country_code

default NULL. An ISO 3166 country code. See iso_3166_codes() for valid ISO codes. Optional.

preferred_label_values

default NULL. Must be one of "postalCity" or "localCity". Optional.

geocoder

default default_geocoder().

token

an object of class httr2_token as generated by auth_code() or related function

Value

A data.frame with 3 columns: text, magic_key, and is_collection.

Details

Unlike the other functions in this package, suggest_places() is not vectorized as it is intended to provide search suggestions for individual queries such as those made in a search bar.

See REST API documentation for more.

Examples

# identify a search point
location <- sf::st_sfc(sf::st_point(c(-84.34, 33.74)), crs = 4326)

# create a search extent from it
search_extent <- sf::st_bbox(sf::st_buffer(location, 10))

# find suggestions from it
suggestions <- suggest_places(
  "bellwood",
  location,
  search_extent = search_extent
)

# get address candidate information
# using the text and the magic key
find_address_candidates(
  suggestions$text,
  magic_key = suggestions$magic_key
)
#> Simple feature collection with 7 features and 60 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -84.34273 ymin: 33.74034 xmax: -84.34273 ymax: 33.74034
#> Geodetic CRS:  WGS 84
#>   input_id loc_name status score                 match_addr
#> 1        1     <NA>      M   100            Bellwood Coffee
#> 2        2     <NA>      M   100 Bellwood, Atlanta, Georgia
#> 3        3     <NA>      M   100            Bellwood Church
#> 4        4     <NA>      M   100              Bellwood Yard
#> 5        5     <NA>      T   100         Bellwood, Illinois
#> 6        5     <NA>      T   100         Bellwood, Illinois
#> 7        5     <NA>      T   100         Bellwood, Illinois
#>                                                       long_label
#> 1 Bellwood Coffee, 1366 Glenwood Ave SE, Atlanta, GA, 30316, USA
#> 2                                     Bellwood, Atlanta, GA, USA
#> 3                              Bellwood Church, Atlanta, GA, USA
#> 4                                Bellwood Yard, Atlanta, GA, USA
#> 5                                              Bellwood, IL, USA
#> 6                                              Bellwood, IL, USA
#> 7                                              Bellwood, IL, USA
#>       short_label addr_type type_field      place_name
#> 1 Bellwood Coffee       POI     Snacks Bellwood Coffee
#> 2        Bellwood  Locality       City        Bellwood
#> 3 Bellwood Church       POI     Church Bellwood Church
#> 4   Bellwood Yard       POI   Building   Bellwood Yard
#> 5        Bellwood  Locality       City        Bellwood
#> 6        Bellwood  Locality       City        Bellwood
#> 7        Bellwood  Locality       City        Bellwood
#>                                      place_addr phone  url  rank add_bldg
#> 1 1366 Glenwood Ave SE, Atlanta, Georgia, 30316  <NA> <NA> 19.00     <NA>
#> 2                              Atlanta, Georgia  <NA> <NA> 20.00     <NA>
#> 3                              Atlanta, Georgia  <NA> <NA> 24.00     <NA>
#> 4                              Atlanta, Georgia  <NA> <NA> 24.00     <NA>
#> 5                            Bellwood, Illinois  <NA> <NA> 10.50     <NA>
#> 6                            Bellwood, Illinois  <NA> <NA> 16.04     <NA>
#> 7                 Village of Bellwood, Illinois  <NA> <NA> 21.00     <NA>
#>   add_num add_num_from add_num_to add_range side st_pre_dir st_pre_type
#> 1    1366         <NA>       <NA>      <NA> <NA>       <NA>        <NA>
#> 2    <NA>         <NA>       <NA>      <NA> <NA>       <NA>        <NA>
#> 3    <NA>         <NA>       <NA>      <NA> <NA>       <NA>        <NA>
#> 4    <NA>         <NA>       <NA>      <NA> <NA>       <NA>        <NA>
#> 5    <NA>         <NA>       <NA>      <NA> <NA>       <NA>        <NA>
#> 6    <NA>         <NA>       <NA>      <NA> <NA>       <NA>        <NA>
#> 7    <NA>         <NA>       <NA>      <NA> <NA>       <NA>        <NA>
#>    st_name st_type st_dir bldg_type bldg_name level_type level_name unit_type
#> 1 Glenwood     Ave     SE      <NA>      <NA>       <NA>       <NA>      <NA>
#> 2     <NA>    <NA>   <NA>      <NA>      <NA>       <NA>       <NA>      <NA>
#> 3     <NA>    <NA>   <NA>      <NA>      <NA>       <NA>       <NA>      <NA>
#> 4     <NA>    <NA>   <NA>      <NA>      <NA>       <NA>       <NA>      <NA>
#> 5     <NA>    <NA>   <NA>      <NA>      <NA>       <NA>       <NA>      <NA>
#> 6     <NA>    <NA>   <NA>      <NA>      <NA>       <NA>       <NA>      <NA>
#> 7     <NA>    <NA>   <NA>      <NA>      <NA>       <NA>       <NA>      <NA>
#>   unit_name sub_addr              st_addr block sector nbrhd district
#> 1      <NA>     <NA> 1366 Glenwood Ave SE  <NA>   <NA>  <NA>     <NA>
#> 2      <NA>     <NA>                 <NA>  <NA>   <NA>  <NA> Bellwood
#> 3      <NA>     <NA>                 <NA>  <NA>   <NA>  <NA>     <NA>
#> 4      <NA>     <NA>                 <NA>  <NA>   <NA>  <NA>     <NA>
#> 5      <NA>     <NA>                 <NA>  <NA>   <NA>  <NA>     <NA>
#> 6      <NA>     <NA>                 <NA>  <NA>   <NA>  <NA>     <NA>
#> 7      <NA>     <NA>                 <NA>  <NA>   <NA>  <NA>     <NA>
#>                  city metro_area     subregion   region region_abbr territory
#> 1             Atlanta       <NA> DeKalb County  Georgia          GA      <NA>
#> 2             Atlanta       <NA> Fulton County  Georgia          GA      <NA>
#> 3             Atlanta       <NA> Fulton County  Georgia          GA      <NA>
#> 4             Atlanta       <NA> Fulton County  Georgia          GA      <NA>
#> 5            Bellwood       <NA>   Cook County Illinois          IL      <NA>
#> 6            Bellwood       <NA>   Cook County Illinois          IL      <NA>
#> 7 Village of Bellwood       <NA>   Cook County Illinois          IL      <NA>
#>   zone postal postal_ext country    cntry_name lang_code distance         x
#> 1 <NA>  30316       <NA>     USA United States       ENG      0.0 -84.34273
#> 2 <NA>   <NA>       <NA>     USA United States       ENG      0.0 -84.41243
#> 3 <NA>   <NA>       <NA>     USA United States       ENG      0.0 -84.41521
#> 4 <NA>   <NA>       <NA>     USA United States       ENG      0.0 -84.41798
#> 5 <NA>   <NA>       <NA>     USA United States       ENG 956972.0 -87.87345
#> 6 <NA>   <NA>       <NA>     USA United States       ENG 957055.1 -87.88312
#> 7 <NA>   <NA>       <NA>     USA United States       ENG 957012.3 -87.87617
#>          y display_x display_y      xmin      xmax     ymin     ymax ex_info
#> 1 33.74034 -84.34273  33.74034 -84.34373 -84.34173 33.73934 33.74134    <NA>
#> 2 33.77455 -84.41243  33.77455 -84.42343 -84.40143 33.76355 33.78555    <NA>
#> 3 33.77288 -84.41521  33.77288 -84.42021 -84.41021 33.76788 33.77788    <NA>
#> 4 33.77927 -84.41798  33.77927 -84.42298 -84.41298 33.77427 33.78427    <NA>
#> 5 41.88325 -87.87345  41.88325 -87.88945 -87.85745 41.86725 41.89925    <NA>
#> 6 41.88142 -87.88312  41.88142 -87.90012 -87.86612 41.86442 41.89842    <NA>
#> 7 41.88290 -87.87617  41.88290 -87.89317 -87.85917 41.86590 41.89990    <NA>
#>                                    extents                   geometry
#> 1 -84.34373, 33.73934, -84.34173, 33.74134 POINT (-84.34273 33.74034)
#> 2 -84.42343, 33.76355, -84.40143, 33.78555 POINT (-84.41243 33.77455)
#> 3 -84.42021, 33.76788, -84.41021, 33.77788 POINT (-84.41521 33.77288)
#> 4 -84.42298, 33.77427, -84.41298, 33.78427 POINT (-84.41798 33.77927)
#> 5 -87.88945, 41.86725, -87.85745, 41.89925 POINT (-87.87345 41.88325)
#> 6 -87.90012, 41.86442, -87.86612, 41.89842 POINT (-87.88312 41.88142)
#> 7 -87.89317, 41.86590, -87.85917, 41.89990  POINT (-87.87617 41.8829)