Search Suggestion

This function returns candidate locations based on a partial search query. It is designed to be used in an interactive search experience in a client facing application.

Usage

suggest_places(
  text,
  location = NULL,
  category = NULL,
  search_extent = NULL,
  max_suggestions = NULL,
  country_code = NULL,
  preferred_label_values = NULL,
  geocoder = default_geocoder(),
  token = arc_token()
)

Arguments

text

a scalar character of search key to generate a place suggestion.

location

an sfc_POINT object that centers the search. Optional.

category

a scalar character. Place or address type that can be used to filter suggest results. Optional.

search_extent

an object of class bbox that limits the search area. This is especially useful for applications in which a user will search for places and addresses within the current map extent. Optional.

max_suggestions

default NULL. The maximum number of suggestions to return. The service default is 5 with a maximum of 15.

country_code

default NULL. An ISO 3166 country code. See iso_3166_codes() for valid ISO codes. Optional.

preferred_label_values

default NULL. Must be one of "postalCity" or "localCity". Optional.

geocoder

default default_geocoder().

token

an object of class httr2_token as generated by auth_code() or related function

Value

A data.frame with 3 columns: text, magic_key, and is_collection.

Details

Unlike the other functions in this package, suggest_places() is not vectorized as it is intended to provide search suggestions for individual queries such as those made in a search bar.

Utilizes the /suggest endpoint.

Examples

# identify a search point
location <- sf::st_sfc(sf::st_point(c(-84.34, 33.74)), crs = 4326)

# create a search extent from it
search_extent <- sf::st_bbox(sf::st_buffer(location, 10))

# find suggestions from it
suggestions <- suggest_places(
  "bellwood",
  location,
  search_extent = search_extent
)

# get address candidate information
# using the text and the magic key
find_address_candidates(
  suggestions$text,
  magic_key = suggestions$magic_key
)
#> # A data frame: 7 × 61
#>   input_id loc_name status score match_addr     long_label short_label addr_type
#> *    <int> <chr>    <chr>  <dbl> <chr>          <chr>      <chr>       <chr>    
#> 1        1 NA       M        100 Bellwood Coff… Bellwood … Bellwood C… POI      
#> 2        2 NA       M        100 Bellwood, Atl… Bellwood,… Bellwood    Locality 
#> 3        3 NA       M        100 Bellwood Chur… Bellwood … Bellwood C… POI      
#> 4        4 NA       M        100 Bellwood Yard  Bellwood … Bellwood Y… POI      
#> 5        5 NA       T        100 Bellwood, Ill… Bellwood,… Bellwood    Locality 
#> 6        5 NA       T        100 Bellwood, Ill… Bellwood,… Bellwood    Locality 
#> 7        5 NA       T        100 Bellwood, Ill… Bellwood,… Bellwood    Locality 
#> # ℹ 53 more variables: type_field <chr>, place_name <chr>, place_addr <chr>,
#> #   phone <chr>, url <chr>, rank <dbl>, add_bldg <chr>, add_num <chr>,
#> #   add_num_from <chr>, add_num_to <chr>, add_range <chr>, side <chr>,
#> #   st_pre_dir <chr>, st_pre_type <chr>, st_name <chr>, st_type <chr>,
#> #   st_dir <chr>, bldg_type <chr>, bldg_name <chr>, level_type <chr>,
#> #   level_name <chr>, unit_type <chr>, unit_name <chr>, sub_addr <chr>,
#> #   st_addr <chr>, block <chr>, sector <chr>, nbrhd <chr>, district <chr>, …